feat(store/products): switch addProduct to products reducer
