+++ client/store/products.js
@@ -4,6 +4,7 @@ import axios from "axios";
  * ACTION TYPES
  */
 const GET_PRODUCTS = "GET_PRODUCTS";
+const ADD_SINGLE_PRODUCT = "ADD_SINGLE_PRODUCT";
 
 /**
  * INITIAL STATE
@@ -14,6 +15,11 @@ const products = [];
  * ACTION CREATORS
  */
 const gotProducts = products => ({ type: GET_PRODUCTS, products });
+
+export const addSingleProduct = product => ({
+	type: ADD_SINGLE_PRODUCT,
+	product
+});
 
 /**
  * THUNK CREATORS
@@ -27,6 +33,17 @@ export const getProducts = () => async dispatch => {
 	}
 };
 
+export const createSingleProduct = product => {
+	return async dispatch => {
+		try {
+			const { data: created } = await axios.post("/api/products/add", product);
+			dispatch(addSingleProduct(created));
+		} catch (err) {
+			console.log("We're having trouble adding this product.");
+		}
+	};
+};
+
 /**
  * REDUCER
  */
@@ -34,6 +51,8 @@ export default function (state = products, action) {
 	switch (action.type) {
 		case GET_PRODUCTS:
 			return action.products;
+		case ADD_SINGLE_PRODUCT:
+			return [...state, action.product];
 		default:
 			return state;
 	}
