+++ client/store/singleProduct.js
@@ -2,16 +2,10 @@ import axios from "axios";
 
 // ACTION TYPES
 const SET_SINGLE_PRODUCT = "SET_SINGLE_PRODUCT";
-const ADD_SINGLE_PRODUCT = "ADD_SINGLE_PRODUCT";
 
 // ACTION CREATORS
 export const setSingleProduct = product => ({
 	type: SET_SINGLE_PRODUCT,
-	product
-});
-
-export const addSingleProduct = product => ({
-	type: ADD_SINGLE_PRODUCT,
 	product
 });
 
@@ -27,19 +21,6 @@ export const fetchSingleProduct = id => {
 	};
 };
 
-export const createSingleProduct = (product, history) => {
-	return async dispatch => {
-		try {
-			console.log("product -->", product);
-			const { data: created } = await axios.post("/api/products/add", product);
-			dispatch(addSingleProduct(created));
-			history.push("/products");
-		} catch (err) {
-			console.log("We're having trouble adding this product.");
-		}
-	};
-};
-
 // INITIAL STATE
 const initialState = {};
 
@@ -47,8 +28,6 @@ const initialState = {};
 export default function singleProductReducer(state = initialState, action) {
 	switch (action.type) {
 		case SET_SINGLE_PRODUCT:
-			return action.product;
-		case ADD_SINGLE_PRODUCT:
 			return action.product;
 		default:
 			return state;
